requirements.txt
# Core Deep Learning
tensorflow>=2.10.0
keras>=2.10.0
torch>=2.0.0
numpy>=1.23.0
pandas>=1.5.0

# Music Processing
music21>=8.1.0
mido>=1.2.10
pretty_midi>=0.2.9

# Data Processing
scikit-learn>=1.2.0
scipy>=1.10.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.13.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
jupyter>=1.0.0
notebook>=6.5.0

# Evaluation
librosa>=0.10.0
mir_eval>=0.7
